images changed and others

+++ services/lleidahackergroup.py
@@ -1,22 +1,39 @@
 from models.LleidaHacker import LleidaHackerGroup as ModelLleidaHackerGroup
 from models.LleidaHacker import LleidaHacker as ModelLleidaHacker
+from models.TokenData import TokenData
+from models.UserType import UserType
 
 from schemas.LleidaHacker import LleidaHackerGroup as SchemaLleidaHackerGroup
+from schemas.LleidaHacker import LleidaHackerGroupUpdate as SchemaLleidaHackerGroupUpdate
+
+from utils.service_utils import set_existing_data
 
 from sqlalchemy.orm import Session
 
+from errors.AuthenticationException import AuthenticationException
+from errors.NotFoundException import NotFoundException
+from errors.InvalidDataException import InvalidDataException
 
 async def get_all(db: Session):
     return db.query(ModelLleidaHackerGroup).all()
 
 
 async def get_lleidahackergroup(groupId: int, db: Session):
-    return db.query(ModelLleidaHackerGroup).filter(
-        ModelLleidaHackerGroup.id == groupId).first()
+    group = db.query(ModelLleidaHackerGroup).filter(ModelLleidaHackerGroup.id == groupId).first()
+    if group is None:
+        raise NotFoundException("LleidaHacker group not found")
+    return group
 
 
-async def add_lleidahackergroup(payload: SchemaLleidaHackerGroup, db: Session):
-    new_lleidahacker_group = ModelLleidaHackerGroup(**payload.dict())
+async def add_lleidahackergroup(payload: SchemaLleidaHackerGroup, db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
+    hacker = db.query(ModelLleidaHacker).filter(
+        ModelLleidaHacker.id == data.user_id).first()
+    if hacker is None:
+        raise NotFoundException("LleidaHacker not found")
+    new_lleidahacker_group = ModelLleidaHackerGroup(**payload.dict(), leader_id=hacker.id)
     db.add(new_lleidahacker_group)
     db.commit()
     db.refresh(new_lleidahacker_group)
@@ -24,31 +41,53 @@ async def add_lleidahackergroup(payload: SchemaLleidaHackerGroup, db: Session):
 
 
 async def update_lleidahackergroup(groupId: int,
-                                   payload: SchemaLleidaHackerGroup,
-                                   db: Session):
+                                   payload: SchemaLleidaHackerGroupUpdate,
+                                   db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
     lleidahacker_group = db.query(ModelLleidaHackerGroup).filter(
         ModelLleidaHackerGroup.id == groupId).first()
-    lleidahacker_group.name = payload.name
-    lleidahacker_group.description = payload.description
+    if lleidahacker_group is None:
+        raise NotFoundException("LleidaHacker group not found")
+    if not (data.user_type == UserType.LLEIDAHACKER.value and data.user_id == lleidahacker_group.leader_id):
+        raise AuthenticationException("Not authorized")
+    updated = set_existing_data(lleidahacker_group, payload)
     db.commit()
     db.refresh(lleidahacker_group)
-    return lleidahacker_group
+    return lleidahacker_group, updated
 
 
-async def delete_lleidahackergroup(groupId: int, db: Session):
+async def delete_lleidahackergroup(groupId: int, db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
     lleidahacker_group = db.query(ModelLleidaHackerGroup).filter(
         ModelLleidaHackerGroup.id == groupId).first()
+    if lleidahacker_group is None:
+        raise NotFoundException("LleidaHacker group not found")
+    if not (data.user_type == UserType.LLEIDAHACKER.value and data.user_id == lleidahacker_group.leader_id):
+        raise AuthenticationException("Not authorized")
     db.delete(lleidahacker_group)
     db.commit()
     return lleidahacker_group
 
 
 async def add_lleidahacker_to_group(groupId: int, lleidahackerId: int,
-                                    db: Session):
+                                    db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
     lleidahacker_group = db.query(ModelLleidaHackerGroup).filter(
         ModelLleidaHackerGroup.id == groupId).first()
+    if lleidahacker_group is None:
+        raise NotFoundException("LleidaHacker group not found")
     lleidahacker = db.query(ModelLleidaHacker).filter(
         ModelLleidaHacker.id == lleidahackerId).first()
+    if not (data.user_type == UserType.LLEIDAHACKER.value and data.user_id == lleidahacker_group.leader_id):
+        raise AuthenticationException("Not authorized")
+    if lleidahacker is None:
+        raise NotFoundException("LleidaHacker not found")
     lleidahacker_group.members.append(lleidahacker)
     db.commit()
     db.refresh(lleidahacker_group)
@@ -56,11 +95,22 @@ async def add_lleidahacker_to_group(groupId: int, lleidahackerId: int,
 
 
 async def remove_lleidahacker_from_group(groupId: int, lleidahackerId: int,
-                                         db: Session):
+                                         db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
     lleidahacker_group = db.query(ModelLleidaHackerGroup).filter(
         ModelLleidaHackerGroup.id == groupId).first()
+    if lleidahacker_group is None:
+        raise NotFoundException("LleidaHacker group not found")
+    if not (data.user_type == UserType.LLEIDAHACKER.value and data.user_id == lleidahacker_group.leader_id):
+        raise AuthenticationException("Not authorized")
     lleidahacker = db.query(ModelLleidaHacker).filter(
         ModelLleidaHacker.id == lleidahackerId).first()
+    if lleidahacker is None:
+        raise NotFoundException("LleidaHacker not found")
+    if lleidahacker not in lleidahacker_group.members:
+        raise InvalidDataException("LleidaHacker not in group")
     lleidahacker_group.members.remove(lleidahacker)
     db.commit()
     db.refresh(lleidahacker_group)
@@ -68,11 +118,22 @@ async def remove_lleidahacker_from_group(groupId: int, lleidahackerId: int,
 
 
 async def set_lleidahacker_group_leader(groupId: int, lleidahackerId: int,
-                                        db: Session):
+                                        db: Session, data: TokenData):
+    if not data.is_admin:
+        if not (data.available and data.user_type == UserType.LLEIDAHACKER.value):
+            raise AuthenticationException("Not authorized")
     lleidahacker_group = db.query(ModelLleidaHackerGroup).filter(
         ModelLleidaHackerGroup.id == groupId).first()
+    if lleidahacker_group is None:
+        raise NotFoundException("LleidaHacker group not found")
+    if not (data.user_type == UserType.LLEIDAHACKER.value and data.user_id == lleidahacker_group.leader_id):
+        raise AuthenticationException("Not authorized")
     lleidahacker = db.query(ModelLleidaHacker).filter(
         ModelLleidaHacker.id == lleidahackerId).first()
+    if lleidahacker is None:
+        raise NotFoundException("LleidaHacker not found")
+    if lleidahacker not in lleidahacker_group.members:
+        raise InvalidDataException("LleidaHacker not in group")
     lleidahacker_group.leader_id = lleidahacker.id
     lleidahacker_group.leader = lleidahacker
